Fix type of cluster
cluster is actually an instance of the private Cluster type, which
extends EventEmitter.  Therefore, annotate cluster as being of type
EventEmitter.

Note:  This code is now not runnable, since assigning to cluster would
require either reference to the events namespace (which isn't imported)
or leaving it null which would fail for subsequent property accesses.
But that's fine, since externs aren't run.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>