Major reworking of streams
Update the annotations to match their definitions in node 0.10.  Also,
stream (the namespace) is not a class, use stream.Stream instead as
appropriate.

This commit doesn't fix the multiple inheritance/mixin issue with Duplex
needing to extend from both Readable and Writable.  This may require
defining some interfaces which can be implemented.

This commit also doesn't touch ReadableStream and WritableStream.  I'm
not sure of the purpose of these classes and since they are heavily used
in other namespaces, I'll leave them alone for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>